app/requirements.txt
asgiref==3.8.1
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
Django==5.1.7
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
python-dotenv==1.0.1
redis==5.2.0
requests==2.32.3
spotipy==2.24.0
sqlparse==0.5.3
urllib3==2.2.3
Werkzeug==3.1.3
